# docs/swagger.yaml
definitions:
...
  types.PodListResponse:
    properties:
      age:
        type: string
      clusterName:
        type: string
      configName:
        type: string
      cpu:
        type: string
      hasUpdated:
        type: boolean
      lastRestartAt:
        type: string
      lastRestartReason:
        type: string
      memory:
        type: string
      name:
        type: string
      namespace:
        type: string
      node:
        type: string
      podIP:
        type: string
      qos:
        type: string
      ready:
        type: string
      restarts:
        type: string
      status:
        type: string
      uid:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/daemonset/{name}/pods:
    get:
      consumes:
      - application/json
      description: Retrieve all pods belonging to a specific daemonset by name with
        namespace as query parameter
      parameters:
      - description: Kubernetes config ID
        in: query
        name: config
        required: true
        type: string
      - description: Cluster name (for multi-cluster configs)
        in: query
        name: cluster
        type: string
      - description: DaemonSet name
        in: path
        name: name
        required: true
        type: string
      - description: Namespace name
        in: query
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: DaemonSet pods
          schema:
            items:
              $ref: '#/definitions/types.PodListResponse'
            type: array
        "400":
          description: Bad request - missing namespace parameter
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: DaemonSet not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - KubeConfig: []
      summary: Get DaemonSet Pods by Name
      tags:
      - Workloads
...
  /api/v1/daemonsets/{namespace}/{name}/pods:
    get:
      consumes:
      - application/json
      description: Retrieve all pods belonging to a specific daemonset by namespace
        and name
      parameters:
      - description: Kubernetes config ID
        in: query
        name: config
        required: true
        type: string
      - description: Cluster name (for multi-cluster configs)
        in: query
        name: cluster
        type: string
      - description: Namespace name
        in: path
        name: namespace
        required: true
        type: string
      - description: DaemonSet name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: DaemonSet pods
          schema:
            items:
              $ref: '#/definitions/types.PodListResponse'
            type: array
        "400":
          description: Bad request - invalid parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: DaemonSet not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - KubeConfig: []
      summary: Get DaemonSet Pods by Namespace and Name
      tags:
      - Workloads
...
  /api/v1/deployment/{name}/pods:
    get:
      consumes:
      - application/json
      description: Retrieve all pods belonging to a specific deployment by name with
        namespace as query parameter
      parameters:
      - description: Kubernetes config ID
        in: query
        name: config
        required: true
        type: string
      - description: Cluster name (for multi-cluster configs)
        in: query
        name: cluster
        type: string
      - description: Deployment name
        in: path
        name: name
        required: true
        type: string
      - description: Namespace name
        in: query
        name: namespace
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: Deployment pods
          schema:
            items:
              type: object
            type: array
        "400":
          description: Bad request - missing namespace parameter
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Deployment not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - KubeConfig: []
      summary: Get Deployment Pods by Name
      tags:
      - Workloads
...
  /api/v1/deployments/{namespace}/{name}/pods:
    get:
      consumes:
      - application/json
      description: Retrieve all pods belonging to a specific deployment by namespace
        and name
      parameters:
      - description: Kubernetes config ID
        in: query
        name: config
        required: true
        type: string
      - description: Cluster name (for multi-cluster configs)
        in: query
        name: cluster
        type: string
      - description: Namespace name
        in: path
        name: namespace
        required: true
        type: string
      - description: Deployment name
        in: path
        name: name
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: Deployment pods
          schema:
            items:
              type: object
            type: array
        "400":
          description: Bad request - invalid parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Deployment not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - KubeConfig: []
      summary: Get Deployment Pods by Namespace and Name
      tags:
      - Workloads
...
  /api/v1/job/{name}/pods:
    get:
      consumes:
      - application/json
      description: Retrieves all pods managed by a specific Job by name with namespace
        as query parameter
      parameters:
      - description: Kubernetes configuration ID
        in: query
        name: config
        required: true
        type: string
      - description: Cluster name for multi-cluster setups
        in: query
        name: cluster
        type: string
      - description: Kubernetes namespace
        in: query
        name: namespace
        required: true
        type: string
      - description: Job name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Job pods
          schema:
            items:
              $ref: '#/definitions/types.PodListResponse'
            type: array
        "400":
          description: Bad request - missing or invalid parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Job not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - KubeConfig: []
      summary: Get Job pods by name
      tags:
      - Workloads
...
  /api/v1/namespaces/{name}/pods:
    get:
      consumes:
      - application/json
      - text/event-stream
      description: Retrieves all pods in a specific namespace with real-time updates
      parameters:
      - description: Kubernetes configuration ID
        in: query
        name: config
        required: true
        type: string
      - description: Cluster name for multi-cluster setups
        in: query
        name: cluster
        type: string
      - description: Namespace name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      - text/event-stream
      responses:
        "200":
          description: Namespace pods
          schema:
            items:
              $ref: '#/definitions/types.PodListResponse'
            type: array
        "400":
          description: Bad request - missing or invalid parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Namespace not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - KubeConfig: []
      summary: Get Namespace pods
      tags:
      - Cluster
...
  /api/v1/nodes/{name}/pods:
    get:
      consumes:
      - application/json
      - text/event-stream
      description: Retrieves all pods running on a specific node with real-time updates
      parameters:
      - description: Kubernetes configuration ID
        in: query
        name: config
        required: true
        type: string
      - description: Cluster name for multi-cluster setups
        in: query
        name: cluster
        type: string
      - description: Node name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      - text/event-stream
      responses:
        "200":
          description: Node pods
          schema:
            items:
              $ref: '#/definitions/types.PodListResponse'
            type: array
        "400":
          description: Bad request - missing or invalid parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Node not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - KubeConfig: []
      summary: Get Node pods
      tags:
      - Cluster
...
  /api/v1/pods:
    get:
      consumes:
      - application/json
      description: Get all pods with real-time updates via Server-Sent Events
      parameters:
      - description: Kubernetes configuration ID
        in: query
        name: config
        required: true
        type: string
      - description: Cluster name
        in: query
        name: cluster
        type: string
      - description: Namespace filter
        in: query
        name: namespace
        type: string
      - description: Node name filter
        in: query
        name: node
        type: string
      - description: Owner type (deployment, daemonset, etc.)
        in: query
        name: owner
        type: string
      - description: Owner name
        in: query
        name: ownerName
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: Stream of pod data
          schema:
            items:
              $ref: '#/definitions/types.PodListResponse'
            type: array
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - KubeConfig: []
      summary: Get Pods (SSE)
      tags:
      - Workloads
...
swagger: "2.0"